feat(photo-upload): integrated changes with dev branch
