+++ javinukai-back/src/main/java/lt/javinukai/javinukai/controller/PhotoController.java
@@ -4,10 +4,10 @@ package lt.javinukai.javinukai.controller;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.RequiredArgsConstructor;
-import lt.javinukai.javinukai.entity.ContestantImageCollection;
+import lt.javinukai.javinukai.entity.PhotoCollection;
 import lt.javinukai.javinukai.enums.ImageSize;
 import lt.javinukai.javinukai.exception.NoImagesException;
-import lt.javinukai.javinukai.service.ImageStorageService;
+import lt.javinukai.javinukai.service.PhotoStorageService;
 import org.springframework.http.MediaType;
 import org.springframework.http.ResponseEntity;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
@@ -22,22 +22,22 @@ import java.util.UUID;
 @Controller
 @RequestMapping("/api/v1/images")
 @RequiredArgsConstructor
-public class ImageController {
-    private final ImageStorageService imageStorageService;
+public class PhotoController {
+    private final PhotoStorageService photoStorageService;
 
 
     @PostMapping(headers = "content-type=multipart/form-data", consumes = "image/jpg")
-    public ResponseEntity<ContestantImageCollection> uploadImages(@RequestParam("image") MultipartFile[] images,
-                                                                  @RequestParam("title") @NotBlank String title,
-                                                                  @RequestParam("description") @NotBlank String description,
-                                                                  @RequestParam("categoryId") @NotNull UUID categoryID,
-                                                                  @AuthenticationPrincipal UserDetails userDetails
+    public ResponseEntity<PhotoCollection> uploadImages(@RequestParam("image") MultipartFile[] images,
+                                                        @RequestParam("title") @NotBlank String title,
+                                                        @RequestParam("description") @NotBlank String description,
+                                                        @RequestParam("categoryId") @NotNull UUID categoryID,
+                                                        @AuthenticationPrincipal UserDetails userDetails
     ) throws IOException {
         if (images.length < 1 ) {
             throw new NoImagesException("No jpg images were provided with request");
         }
         return ResponseEntity.ok()
-                .body(imageStorageService.createImage(images, title, description, categoryID, userDetails));
+                .body(photoStorageService.createImage(images, title, description, categoryID, userDetails));
     }
 
     @GetMapping("{image}")
@@ -45,7 +45,7 @@ public class ImageController {
             throws IOException {
         UUID imageId = UUID.fromString(image.split("\\.")[0]);
         return ResponseEntity.ok().contentType(MediaType.IMAGE_JPEG)
-                .body(imageStorageService.getImageById(imageId, ImageSize.valueOf(size.toUpperCase())));
+                .body(photoStorageService.getImageById(imageId, ImageSize.valueOf(size.toUpperCase())));
     }
 
 }
