--- javinukai-back/src/main/java/lt/javinukai/javinukai/entity/ContestantImage.java
@@ -1,54 +1,0 @@
-package lt.javinukai.javinukai.entity;
-
-import com.fasterxml.jackson.annotation.JsonIgnore;
-import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.Builder;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
-
-import java.time.ZonedDateTime;
-import java.util.UUID;
-
-@Entity
-@Builder
-@Getter
-@NoArgsConstructor
-@AllArgsConstructor
-@Table(name = "images")
-public class ContestantImage {
-
-    @Id
-    private UUID uuid;
-
-    private String localPathFull;
-    private String localPathMiddle;
-    private String localPathSmall;
-    private String localPathThumbnail;
-    private String urlFull;
-    private String urlMiddle;
-    private String urlSmall;
-    private String urlThumbnail;
-
-    @JsonIgnore
-    @ManyToOne(cascade = CascadeType.MERGE)
-    @JoinColumn(name = "collection_uuid")
-    private ContestantImageCollection collection;
-
-    @JsonIgnore
-    @ManyToOne(cascade = CascadeType.MERGE)
-    @JoinColumn(name = "author_uuid")
-    private User author;
-
-    private ZonedDateTime createdAt;
-    private ZonedDateTime modifiedAt;
-
-    @PrePersist
-    protected void onCreate() {
-        this.createdAt = ZonedDateTime.now();
-    }
-    @PreUpdate
-    protected void onUpdate() {
-        this.modifiedAt = ZonedDateTime.now();
-    }
-}
