+++ javinukai-back/src/main/java/lt/javinukai/javinukai/repository/PhotoCollectionRepository.java
@@ -1,0 +1,9 @@
+package lt.javinukai.javinukai.repository;
+
+import lt.javinukai.javinukai.entity.PhotoCollection;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+import java.util.UUID;
+
+public interface PhotoCollectionRepository extends JpaRepository<PhotoCollection, UUID> {
+}
