--- javinukai-back/src/main/java/lt/javinukai/javinukai/service/PhotoService.java
@@ -1,14 +1,0 @@
-package lt.javinukai.javinukai.service;
-
-import lt.javinukai.javinukai.entity.Photo;
-import lt.javinukai.javinukai.repository.PhotoRepository;
-
-public class PhotoService {
-    private final PhotoRepository photoRepository;
-
-    public PhotoService(PhotoRepository photoRepository) {
-        this.photoRepository = photoRepository;
-    }
-
-
-}
